% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{bigmem}
\alias{bigmem}
\title{Bigmem SLURM Resource Configuration}
\format{
An object of class \code{list} of length 1.
}
\usage{
bigmem
}
\description{
Establishes SLURM resources for tasks requiring significant memory via
\code{targets::tar_resources} with a "bigmem" controller. Suitable for memory-intensive jobs.
}
\details{
\tabular{lll}{
   CPUs \tab Memory (GB) \tab Time (minutes) \cr
   10 \tab 500 \tab 360 \cr
}
}
\seealso{
\code{\link[targets]{tar_resources}}, \code{\link[targets]{tar_resources_crew}}
}
\keyword{datasets}
