% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_explorer.R
\name{clean_phase_name}
\alias{clean_phase_name}
\title{Remove hash suffix from phase name}
\usage{
clean_phase_name(phase)
}
\arguments{
\item{phase}{Character vector of phase names with IDs}
}
\value{
Character vector of cleaned phase names
}
\description{
Remove hash suffix from phase name
}
