% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_explorer.R
\name{find_targets_file}
\alias{find_targets_file}
\title{Find _targets.R file from logs path}
\usage{
find_targets_file(logs_path)
}
\arguments{
\item{logs_path}{Path to the logs directory}
}
\value{
Path to _targets.R file or NULL if not found
}
\description{
Find _targets.R file from logs path
}
\keyword{internal}
