% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cluster}
\alias{get_cluster}
\title{Determine Cluster Based on Hostname}
\usage{
get_cluster()
}
\value{
A character string: either "apollo" or "gemini".
}
\description{
Retrieves the name of the COH HPRCC cluster by matching the system's hostname
against known patterns. It supports 'apollo' and 'gemini' clusters.
}
\note{
This function will throw an error if the hostname does not match any known cluster pattern.
}
\examples{
if (interactive()) {
    get_cluster()
}
}
