% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{huge}
\alias{huge}
\title{Huge SLURM Resource Configuration}
\format{
An object of class \code{list} of length 1.
}
\usage{
huge
}
\description{
Sets up SLURM resources for very large tasks using \code{targets::tar_resources}
with a "huge" controller. Designed for very high computational loads.
}
\details{
\tabular{lll}{
   CPUs \tab Memory (GB) \tab Time (minutes) \cr
   40 \tab 100 \tab 120 \cr
}
}
\seealso{
\code{\link[targets]{tar_resources}}, \code{\link[targets]{tar_resources_crew}}
}
\keyword{datasets}
