% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_job_result.R
\name{is.slurm_job_result}
\alias{is.slurm_job_result}
\title{Test if Object is a slurm_job_result}
\usage{
is.slurm_job_result(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Test if Object is a slurm_job_result
}
