% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_resource_shortcuts.R
\docType{data}
\name{large}
\alias{large}
\title{Large SLURM Resource Configuration}
\format{
An object of class \code{list} of length 1.
}
\usage{
large
}
\description{
Configures SLURM resources for large-scale tasks using \code{targets::tar_resources}
with a "large" controller setting. Ideal for high-demand computations.
}
\details{
\tabular{lll}{
   CPUs \tab Memory (GB) \tab Time (minutes) \cr
   20 \tab 200 \tab 720 \cr
}
}
\seealso{
\code{\link[targets]{tar_resources}}, \code{\link[targets]{tar_resources_crew}}
}
\keyword{datasets}
