% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_message}
\alias{log_message}
\title{Print a message and optionally write it to a log file}
\usage{
log_message(message_str, log = TRUE, log_dir = "logs")
}
\arguments{
\item{message_str}{The message to be logged.}

\item{log}{Logical indicating whether to write the message to a log file. Default is TRUE.}

\item{log_dir}{The directory to store the log file. Default is "logs".}
}
\value{
None, prints a message to the terminal
}
\description{
This function prints a message and optionally writes it to a log file.
}
\examples{
if (interactive()) {
    log_message("This is a log message")
}
}
