% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{medium}
\alias{medium}
\title{Medium SLURM Resource Configuration}
\format{
An object of class \code{list} of length 1.
}
\usage{
medium
}
\description{
Configures medium-level SLURM resources. This object is constructed using
\code{targets::tar_resources} with a "medium" controller setting, suitable for
moderately demanding computational tasks.
}
\details{
\tabular{lll}{
   CPUs \tab Memory (GB) \tab Time (minutes) \cr
   12 \tab 80 \tab 360 \cr
}
}
\seealso{
\code{\link[targets]{tar_resources}}, \code{\link[targets]{tar_resources_crew}}
}
\keyword{datasets}
