% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_analysis.R
\name{plot_target_logs}
\alias{plot_target_logs}
\title{Plot combined metrics from multiple log files}
\usage{
plot_target_logs(logs, metric = "resident", job_ids = NULL, ...)
}
\arguments{
\item{logs}{Data frame returned by read_target_logs()}

\item{metric}{Character string with the name of metric to plot}

\item{job_ids}{Optional character vector of SLURM job IDs to include}

\item{...}{Additional arguments passed to autometric::log_plot()}
}
\value{
A base plot of the metric over time, colored by SLURM job ID
}
\description{
Plot combined metrics from multiple log files
}
\examples{
\dontrun{
logs <- read_target_logs()
plot_target_logs(logs, metric = "rss")
}
}
