% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{small}
\alias{small}
\title{Small SLURM Resource Configuration}
\format{
An object of class \code{list} of length 1.
}
\usage{
small
}
\description{
Defines SLURM resources for small-scale computations. Utilizes \code{targets::tar_resources}
with a "small" controller setting, suitable for less resource-intensive tasks.
}
\details{
\tabular{lll}{
   CPUs \tab Memory (GB) \tab Time (minutes) \cr
   2 \tab 20 \tab 360 \cr
}
}
\seealso{
\code{\link[targets]{tar_resources}}, \code{\link[targets]{tar_resources_crew}}
}
\keyword{datasets}
