% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_resource_shortcuts.R
\docType{data}
\name{tiny}
\alias{tiny}
\title{Tiny SLURM Resource Configuration}
\format{
An object of class \code{list} of length 1.
}
\usage{
tiny
}
\description{
Defines SLURM resources for testing or very small-scale computations. Utilizes \code{targets::tar_resources}
with a "tiny" controller setting, suitable for less resource-intensive tasks.
}
\details{
\tabular{lll}{
   CPUs \tab Memory (GB) \tab Time (minutes) \cr
   1 \tab 1 \tab 60 \cr
}
}
\seealso{
\code{\link[targets]{tar_resources}}, \code{\link[targets]{tar_resources_crew}}
}
\keyword{datasets}
