% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_controller}
\alias{create_controller}
\title{Set Up a Controller for SLURM Jobs on COH Clusters}
\usage{
create_controller(
  name,
  slurm_cpus,
  slurm_mem_gigabytes,
  slurm_walltime_minutes = 720L,
  slurm_workers = 350L,
  slurm_partition = default_partition()
)
}
\arguments{
\item{name}{A unique identifier for the controller.}

\item{slurm_cpus}{Number of CPU cores allocated to each task.}

\item{slurm_mem_gigabytes}{Memory allocated to each task, in gigabytes.}

\item{slurm_walltime_minutes}{Maximum allowed execution time per task, in minutes. Defaults to 720 (12 hours).}

\item{slurm_workers}{Total number of parallel tasks the controller can handle. Defaults to 350.}

\item{slurm_partition}{SLURM partition for job submission. Default set by cluster.
See \link[=package-options]{package options} for defaults.}
}
\value{
A \code{crew_controller} object, ready to manage SLURM job submissions and monitoring.
}
\description{
Configures and initializes a \link[crew.cluster:crew_controller_slurm]{controller}
for managing SLURM jobs on City of Hope clusters using the \link[crew.cluster:crew.cluster-package]{crew.cluster}
package to facilitate job execution, managing resources such as CPU, memory, walltime, and
writing SLURM logs and scripts.
}
\details{
\code{create_controller} streamlines SLURM job setup on COH clusters using
Singularity containers for consistent computing environments. Singularity containers
package software and dependencies, ensuring that jobs run reliably across both
clusters. This approach aids in computational reproducibility and
solves environment inconsistencies between Apollo and Gemini.

The function allows customization of SLURM job parameters, including CPUs, memory,
and walltime, while managing SLURM logs and script directories. It abstracts cluster-specific
configurations, making it easier to run jobs without detailed knowledge of the underlying
cluster setup. This functionality is especially useful in environments with varying
resource paths and setups, simplifying job execution across the two platforms.
}
\examples{
\dontrun{
 # Basic controller with minimal resources
 ctrl <- create_controller("test",
                        slurm_cpus = 2,
                        slurm_mem_gigabytes = 8)
}
# GPU configuration on Gemini
\dontrun{
 if (get_cluster() == "gemini") {
      gpu_ctrl <- create_controller("gpu_job",
                              slurm_cpus = 4,
                              slurm_mem_gigabytes = 60,
                              slurm_partition = "gpu-a100")
 }
}

# Retry controller with escalating resources
\dontrun{
retry_ctrl <- create_controller("retry",
                              slurm_cpus = c(2, 4, 8),
                              slurm_mem_gigabytes = c(8, 16, 32))
}
}
\seealso{
\code{\link[crew.cluster]{crew_controller_slurm}} for more on SLURM controllers.
}
