% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_job.R
\name{generate_slurm_script}
\alias{generate_slurm_script}
\title{Generate SLURM script content}
\usage{
generate_slurm_script(
  name,
  command,
  working_dir,
  slurm_options,
  modules_to_load,
  env_vars
)
}
\arguments{
\item{name}{Job name}

\item{command}{Shell command to run}

\item{working_dir}{Working directory}

\item{slurm_options}{Named list of SLURM options}

\item{modules_to_load}{Character vector of modules}

\item{env_vars}{Named list of environment variables}
}
\value{
Character string with full script content
}
\description{
Generate SLURM script content
}
\keyword{internal}
