% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{package-options}
\alias{package-options}
\title{Package Options}
\description{
This page documents the configurable options available in the \strong{hprcc} package, settable using \code{\link[base]{options}}.
}
\section{log_slurm}{

Controls whether SLURM job scripts logging is turned on or off. \cr
Default: \code{options(hprcc.log_slurm = FALSE)}
}

\section{r_libs_user}{

Sets the path to user R libraries. \cr
Default: \code{options(hprcc.r_libs_user = "~/R/bioc-3.17")}
}

\section{r_libs_site_apollo}{

Sets the site-specific library path for the Apollo cluster. \cr
Default: \code{options(hprcc.r_libs_site_apollo = "/opt/singularity-images/rbioc/rlibs/bioc-3.17")}
}

\section{r_libs_site_gemini}{

Sets the site-specific library path for the Gemini cluster. \cr
Default: \code{options(hprcc.r_libs_site_gemini = "/packages/singularity/shared_cache/rbioc/rlibs/bioc-3.17")}
}

\section{singularity_bin_apollo}{

Sets the path to the Singularity binary on the Apollo cluster. \cr
Default: \code{options(hprcc.singularity_bin_apollo = "/opt/singularity/3.7.0/bin/singularity")}
}

\section{singularity_bin_gemini}{

Sets the path to the Singularity binary on the Gemini cluster. \cr
Default: \code{options(hprcc.singularity_bin_gemini = "/packages/easy-build/software/singularity/3.7.0/bin/singularity")}
}

\section{singularity_container_apollo}{

Sets the path to the Singularity image on the Apollo cluster. \cr
Default: \code{options(hprcc.singularity_container_apollo = "/opt/singularity-images/rbioc/vscode-rbioc_3.17.sif")}
}

\section{singularity_container_gemini}{

Sets the path to the Singularity image on the Gemini cluster. \cr
Default: \code{options(hprcc.singularity_container_gemini = "/packages/singularity/shared_cache/rbioc/vscode-rbioc_3.17.sif")}
}

\section{bind_dirs_apollo}{

Sets the directories to bind in the Singularity container on the Apollo cluster. \cr
Default: \code{options(hprcc.bind_dirs_apollo = "/labs,/opt/singularity,/opt/singularity-images")}
}

\section{bind_dirs_gemini}{

Sets the directories to bind in the Singularity container on the Gemini cluster. \cr
Default: \code{options(hprcc.bind_dirs_gemini = "/packages/singularity,/ref_genomes,/scratch")}
}

\section{default_partition_apollo}{

Sets the default SLURM partition for the Apollo cluster. \cr
Default: \code{options(hprcc.default_partition_apollo = "fast,all")}
}

\section{default_partition_gemini}{

Sets the default SLURM partition for the Gemini cluster. \cr
Default: \code{options(hprcc.default_partition_gemini = "defq")}
}

