% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_resource_shortcuts.R
\docType{data}
\name{SLURM-Resource-Configurations}
\alias{SLURM-Resource-Configurations}
\alias{tiny}
\alias{small}
\alias{medium}
\alias{large}
\alias{large_mem}
\alias{xlarge}
\alias{huge}
\alias{gpu_medium}
\alias{gpu_large}
\title{SLURM Resource Configurations}
\description{
Defines SLURM resources for a variety of computational scales, from testing and small-scale computations
to very large tasks with high computational demands. Utilizes \code{targets::tar_resources}
with specific controller settings suitable for the respective task sizes.
}
\details{
The available configurations are:\tabular{llll}{
   Job Type \tab CPUs \tab Memory (GB) \tab Time (minutes) \cr
   tiny \tab 2 \tab 8 \tab 60 \cr
   small \tab 2 \tab 20 \tab 360 \cr
   medium \tab 4 \tab 40 \tab 360 \cr
   large \tab 8 \tab 80 \tab 360 \cr
   large_mem \tab 8 \tab 800 \tab 360 \cr
   xlarge \tab 20 \tab 200 \tab 360 \cr
   huge \tab 40 \tab 200 \tab 120 \cr
   gpu_medium\eqn{^†}   \tab 4 \tab 60 \tab 120 \cr
   gpu_large\eqn{^†}   \tab 8 \tab 120 \tab 240 \cr
}


\eqn{^†} GPUs only available on \emph{Gemini}
}
\seealso{
\code{\link[targets]{tar_resources}}, \code{\link[targets]{tar_resources_crew}}
}
